move forestry interface to prevent future circular issues
