+++ src/lib/skilling/skills/Woodcutting/woodcutting.ts
@@ -1,8 +1,8 @@
 import { LootTable } from 'oldschooljs';
 
-import { Emoji } from '../../constants';
-import itemID from '../../util/itemID';
-import { Log, SkillsEnum } from '../types';
+import { Emoji } from '../../../constants';
+import itemID from '../../../util/itemID';
+import { Log, SkillsEnum } from '../../types';
 
 const sulliuscepTable = new LootTable()
 	.add('Numulite', [4, 8], 34)
